requirements.txt
Django == 4.1.5
djangorestframework == 3.14.0
djangorestframework-simplejwt == 4.8.0
djoser == 2.1.0
psycopg2 == 2.9.5
psycopg2-binary == 2.9.5
django-cors-headers == 3.13.0